TR-1512 Refactor implementation to move modal outside of popover so it will be visible
